Consolidate Qwen3 runner onto shared VLLM runner
