+++ agents/architectures/vllm_runner.py
@@ -1,67 +1,227 @@
 """
-Placeholder vLLM runner implementation for future integration.
+Unified runner implementation used by vLLM and Qwen3 runner entry points.
+
+This implementation currently uses HuggingFace Transformers for inference while
+keeping the vLLM runner interface stable.
 """
-from typing import Dict, Any, Optional
+from typing import Dict, Any, Optional, List
 import logging
+import os
+import glob
+import json
+import torch
+import transformers
+from transformers import AutoTokenizer, AutoModelForCausalLM, AutoConfig
+from huggingface_hub import login
+import safetensors.torch
 from .base_runner import BaseRunner, GenerationConfig
+from agents.models.llama_completion import strings_to_message_dict
 
 logger = logging.getLogger(__name__)
 
+DEFAULT_QWEN3_MODEL_ID = "Qwen/Qwen3-8B"
+
+
 class VLLMRunner(BaseRunner):
-    """Placeholder runner for vLLM-based inference."""
-    
+    """Runner for vLLM/Qwen3-style model inference via Transformers."""
+
     def __init__(self):
         self.model = None
+        self.tokenizer = None
+        self._pipeline = None
         self.model_id = None
-        
+        self.weights_dir = None
+        self.device = None
+
     def load(self, model_id: str, device_config: Optional[Dict[str, Any]] = None) -> None:
-        """
-        Load the vLLM model (placeholder implementation).
-        
-        Args:
-            model_id: Identifier for the model to load
-            device_config: Optional device configuration
-        """
+        """Load a model for inference from local files or HuggingFace Hub."""
+        device_config = device_config or {}
         self.model_id = model_id
-        logger.warning("vLLM runner is a placeholder implementation")
-        
-        # TODO: Implement actual vLLM model loading
-        # Example:
-        # from vllm import LLM
-        # self.model = LLM(model=model_id, **device_config or {})
-        
-        raise NotImplementedError("vLLM runner not yet implemented")
-    
+
+        self.weights_dir = device_config.get(
+            "weights_dir",
+            os.path.join("app/model_weights", model_id.replace("/", "_"))
+        )
+
+        if "device" in device_config:
+            self.device = torch.device(device_config["device"])
+        elif torch.cuda.is_available():
+            self.device = torch.device("cuda")
+            logger.info("Using CUDA device")
+        elif torch.backends.mps.is_available():
+            self.device = torch.device("mps")
+            logger.info("Using MPS (Apple Silicon) device")
+        else:
+            self.device = torch.device("cpu")
+            logger.info("Using CPU device")
+
+        config_path = os.path.join(self.weights_dir, "config.json")
+        has_config = os.path.exists(config_path)
+        safetensors_files = self._find_safetensors_files(self.weights_dir) if has_config else []
+        has_hf_index = (
+            os.path.exists(os.path.join(self.weights_dir, "model.safetensors.index.json"))
+            or os.path.exists(os.path.join(self.weights_dir, "model.safetensors"))
+        )
+        has_bin = (
+            os.path.exists(os.path.join(self.weights_dir, "pytorch_model.bin"))
+            or os.path.exists(os.path.join(self.weights_dir, "pytorch_model.bin.index.json"))
+        )
+
+        if has_config and (has_hf_index or has_bin):
+            logger.info(f"Loading model from HF pretrained directory: {self.weights_dir}")
+            self._load_from_local()
+        elif has_config and safetensors_files:
+            logger.info(
+                f"Loading model from safetensors files: {self.weights_dir} "
+                f"({len(safetensors_files)} files)"
+            )
+            self._load_from_safetensors(safetensors_files)
+        elif has_config:
+            logger.info(f"Loading model from local directory: {self.weights_dir}")
+            self._load_from_local()
+        else:
+            logger.info(f"Loading model from HuggingFace Hub: {self.model_id}")
+            self._load_from_hub()
+
+        logger.info(f"Runner loaded for model: {model_id}")
+
+    @staticmethod
+    def _find_safetensors_files(directory: str) -> List[str]:
+        return sorted(glob.glob(os.path.join(directory, "*.safetensors")))
+
+    def _load_from_local(self) -> None:
+        self.tokenizer = AutoTokenizer.from_pretrained(self.weights_dir)
+        self.model = AutoModelForCausalLM.from_pretrained(
+            self.weights_dir,
+            torch_dtype=torch.float16,
+        )
+        self.model = self.model.to(self.device)
+        logger.info(f"Model loaded from local weights. Parameters: {self.model.num_parameters()}")
+
+    def _load_from_safetensors(self, safetensors_files: List[str]) -> None:
+        tokenizer_files = [
+            os.path.join(self.weights_dir, f)
+            for f in ("tokenizer.json", "tokenizer_config.json", "tokenizer.model")
+        ]
+        if any(os.path.exists(f) for f in tokenizer_files):
+            self.tokenizer = AutoTokenizer.from_pretrained(self.weights_dir)
+        else:
+            logger.info("Tokenizer not found locally, loading from hub")
+            self.tokenizer = AutoTokenizer.from_pretrained(self.model_id)
+
+        config_path = os.path.join(self.weights_dir, "config.json")
+        with open(config_path, "r") as f:
+            config_json = json.load(f)
+
+        torch_dtype_str = str(config_json.get("torch_dtype", "")).lower()
+        load_dtype = torch.bfloat16 if "bfloat16" in torch_dtype_str else torch.float16
+
+        config = AutoConfig.from_pretrained(self.weights_dir)
+        self.model = AutoModelForCausalLM.from_config(config)
+        self.model = self.model.to(load_dtype)
+
+        state_dict = {}
+        for file_path in safetensors_files:
+            logger.info(f"Loading safetensors file: {file_path}")
+            tensors = safetensors.torch.load_file(file_path, device="cpu")
+            state_dict.update(tensors)
+
+        missing, unexpected = self.model.load_state_dict(state_dict, strict=False)
+        if missing:
+            logger.warning(f"Missing keys when loading safetensors: {len(missing)} keys")
+        if unexpected:
+            logger.warning(f"Unexpected keys when loading safetensors: {len(unexpected)} keys")
+
+        self.model = self.model.to(self.device)
+        logger.info(f"Model loaded from safetensors. Parameters: {self.model.num_parameters()}")
+
+    def _load_from_hub(self) -> None:
+        token = os.environ.get("HUGGING_FACE_HUB_TOKEN")
+        if token:
+            login(token=token)
+
+        self.tokenizer = AutoTokenizer.from_pretrained(self.model_id)
+        self.model = AutoModelForCausalLM.from_pretrained(
+            self.model_id,
+            torch_dtype=torch.float16,
+        )
+        self.model = self.model.to(self.device)
+        logger.info(f"Model loaded from HuggingFace hub. Parameters: {self.model.num_parameters()}")
+
     def generate(self, prompt: str, generation_config: GenerationConfig) -> Dict[str, Any]:
-        """
-        Generate text using vLLM (placeholder).
-        
-        Args:
-            prompt: Input text prompt
-            generation_config: Generation parameters
-            
-        Returns:
-            Dictionary containing generated text and metadata
-        """
-        raise NotImplementedError("vLLM runner not yet implemented")
-    
+        if not self.model or not self.tokenizer:
+            raise RuntimeError("Model not loaded. Call load() first.")
+
+        return self.complete("", prompt, generation_config)
+
     def complete(self, system_prompt: str, user_prompt: str, generation_config: GenerationConfig) -> Dict[str, Any]:
-        """
-        Complete using vLLM (placeholder).
-        
-        Args:
-            system_prompt: System instructions
-            user_prompt: User input
-            generation_config: Generation parameters
-            
-        Returns:
-            Dictionary containing completion and metadata
-        """
-        raise NotImplementedError("vLLM runner not yet implemented")
-    
+        if not self.model or not self.tokenizer:
+            raise RuntimeError("Model not loaded. Call load() first.")
+
+        messages = []
+        if system_prompt:
+            messages.append({"role": "system", "content": system_prompt})
+        messages.append({"role": "user", "content": user_prompt})
+
+        if hasattr(self.tokenizer, "apply_chat_template"):
+            prompt_text = self.tokenizer.apply_chat_template(
+                messages,
+                tokenize=False,
+                add_generation_prompt=True,
+            )
+        else:
+            prompt_text = ""
+            if system_prompt:
+                prompt_text += f"<|im_start|>system\n{system_prompt}<|im_end|>\n"
+            prompt_text += f"<|im_start|>user\n{user_prompt}<|im_end|>\n"
+            prompt_text += "<|im_start|>assistant\n"
+
+        logger.info(
+            f"Generating with temperature={generation_config.temperature}, "
+            f"top_p={generation_config.top_p}, max_tokens={generation_config.max_tokens}"
+        )
+
+        if self._pipeline is None:
+            self._pipeline = transformers.pipeline(
+                "text-generation",
+                model=self.model,
+                tokenizer=self.tokenizer,
+                device=self.device,
+            )
+
+        outputs = self._pipeline(
+            prompt_text,
+            max_new_tokens=generation_config.max_tokens,
+            do_sample=generation_config.temperature > 0,
+            temperature=generation_config.temperature if generation_config.temperature > 0 else None,
+            top_p=generation_config.top_p,
+        )
+
+        output_text = outputs[0]["generated_text"]
+        response_text = output_text[len(prompt_text):] if output_text.startswith(prompt_text) else output_text
+
+        for marker in ["<|im_end|>", "<|im_start|>", "<|endoftext|>"]:
+            idx = response_text.find(marker)
+            if idx != -1:
+                response_text = response_text[:idx]
+
+        response_text = response_text.strip()
+
+        logger.info("Completion successful")
+        logger.info(f"Output: {response_text}")
+
+        return strings_to_message_dict(user_prompt, response_text)
+
     def cleanup(self) -> None:
-        """Clean up vLLM resources (placeholder)."""
+        if self._pipeline:
+            del self._pipeline
+            self._pipeline = None
         if self.model:
-            # TODO: Implement cleanup
+            del self.model
             self.model = None
-            logger.info("vLLM runner cleaned up")
+        if self.tokenizer:
+            del self.tokenizer
+            self.tokenizer = None
+        if torch.cuda.is_available():
+            torch.cuda.empty_cache()
+        logger.info("Runner cleaned up")
